서버에서 응답으로 conversation 받아오기 성공
